anova.glmm.admb {glmmADMB}	R Documentation

Anova for glmm.admb objects

Description

Calculate an anova table from a series of model fits

Usage

anova.glmm.admb(object, ...)

Arguments

object 	 A model fit returned from glmm.admb

... 	 One or more model fits with same random effects

Details

Basic anova table for testing the fixed effects (may be extended later)

Value

An anova table

Author(s)

H. Skaug skaug@mi.uib.no, David Fournier otter@otter-rsch.com and Anders Nielsen andersn@hawaii.edu

Examples

  data(epil2)
  f1<-glmm.admb(y~Base*trt+Age+Visit,random=~Visit,group="subject",data=epil2,family="nbinom")
  f2<-glmm.admb(y~Base*trt+Age,random=~Visit,group="subject",data=epil2,family="nbinom")
  anova(f2,f1)

[Package glmmADMB version 0.3 Index]