glmmADMB-internal {glmmADMB}	R Documentation

Internal glmmADMB Functions

Description

Internal glmmADMB functions

Details

These are not to be called by the user (or in some cases are just waiting for proper documentation to be written :).

[Package glmmADMB version 0.3 Index]