glmmADMB-package {glmmADMB}	R Documentation

Generalized Linear Mixed Models using AD Model Builder

Description

Fits mixed-effects models to count data using Binomial, Poisson or negative binomial response distributions. Zero-inflated versions of Poisson and negative binomial distributions are available.

Details

Package: 	 glmmADMB

Version: 	 0.2

License: 	 GPL

Packaged: 	 Thu Dec 1 07:02:40 2005; andersn

Built: 	 R 2.2.0; ; 2005-12-01 07:13:48; unix

Index:

anova.glmm.admb         Anova for glmm.admb objects
epil2                   Seizure Counts for Epileptics
glmm.admb               Generalized Linear Mixed Models using AD Model Builder
glmmADMB-package        What the package does (this help file)
logLik.glmm.admb        logLik.glmm.admb
plot.glmm.admb          plot.glmm.admb
print.glmm.admb         Print glmm.admb objects
ranef.glmm.admb         Extract Random Effects
summary.glmm.admb       summary.glmm.admb

Author(s)

Hans Skaug <Hans.Skaug@mi.uib.no>, Dave Fournier <otter@otter-rsch.com>, Anders Nielsen <andersn@hawaii.edu>

Maintainer: Hans Skaug <Hans.Skaug@mi.uib.no>

[Package glmmADMB version 0.3 Index]