ranef.glmm.admb {glmmADMB}	R Documentation

Extract Random Effects

Description

Extract random effects from objects fitted with the function 'glmm.admb' and (optionally) their standard deviations.

Usage

ranef.glmm.admb(object, sd = FALSE)

Arguments

object 	 An object (a fit) returned from glmm.admb

sd 	 logical. If 'TRUE', the standard deviations of the random effects are returned, rather than point estimates.

... 	 Further arguments passed on

Details

The point estimates of the random effects are 'empirical Bayes' estimates in the sense that the hyperparameters (fixed effects, variance components etc) are fixed at their MLE's when the joint likelihood is maximized with respect to the random effects. However, uncertainty about the hyperparameters is taken into account when the standard deviations of the random effects are calculated.

Value

A matrix of estimated random effects (or standard deviations).

Author(s)

H. Skaug skaug@mi.uib.no, David Fournier otter@otter-rsch.com and Anders Nielsen andersn@hawaii.edu

See Also

glmm.admb

[Package glmmADMB version 0.3 Index]